requirements.txt
llama-cpp-python
langchain
langchain-community
chromadb
sentence-transformers
thefuzz
huggingface_hub>=0.23
psutil>=5.9
